made project compatible with MonoDevelop 4
* converted project and solution files
* use NUnit instead of Microsoft.VisualStudio.TestTools.UnitTesting
