+++ SharpFileSystem.sln
@@ -1,6 +1,6 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
+# Visual Studio 2012
 VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{0F7EEFA8-7215-406A-9353-849DF790CAA0}"
@@ -13,7 +13,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharpFileSystem.Resources",
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharpFileSystem.SharpZipLib", "SharpFileSystem.SharpZipLib\SharpFileSystem.SharpZipLib.csproj", "{E0BAC37A-07E5-4192-8BD9-51C7CA66C165}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharpFileSystem.Tests", "SharpFileSystem.Tests\SharpFileSystem.Tests.csproj", "{E4BA8FB4-ACD6-43EB-BB03-14AA0C376F5B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharpFileSystem.Tests", "SharpFileSystem.Tests\SharpFileSystem.Tests.csproj", "{5B828500-C591-4B35-AFB2-D38026A7ECBC}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,26 +21,81 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3C4D07BA-23B5-4A63-92CF-F3BF892B7329}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3C4D07BA-23B5-4A63-92CF-F3BF892B7329}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3C4D07BA-23B5-4A63-92CF-F3BF892B7329}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3C4D07BA-23B5-4A63-92CF-F3BF892B7329}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EA7CBA29-CFA8-4945-BB64-CCF25078190E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EA7CBA29-CFA8-4945-BB64-CCF25078190E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EA7CBA29-CFA8-4945-BB64-CCF25078190E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EA7CBA29-CFA8-4945-BB64-CCF25078190E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{34CA60AC-D3A0-4A58-BCE1-3CEA55844715}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{34CA60AC-D3A0-4A58-BCE1-3CEA55844715}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{34CA60AC-D3A0-4A58-BCE1-3CEA55844715}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{34CA60AC-D3A0-4A58-BCE1-3CEA55844715}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3C4D07BA-23B5-4A63-92CF-F3BF892B7329}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3C4D07BA-23B5-4A63-92CF-F3BF892B7329}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3C4D07BA-23B5-4A63-92CF-F3BF892B7329}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3C4D07BA-23B5-4A63-92CF-F3BF892B7329}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5B828500-C591-4B35-AFB2-D38026A7ECBC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5B828500-C591-4B35-AFB2-D38026A7ECBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5B828500-C591-4B35-AFB2-D38026A7ECBC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5B828500-C591-4B35-AFB2-D38026A7ECBC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E0BAC37A-07E5-4192-8BD9-51C7CA66C165}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E0BAC37A-07E5-4192-8BD9-51C7CA66C165}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E0BAC37A-07E5-4192-8BD9-51C7CA66C165}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E0BAC37A-07E5-4192-8BD9-51C7CA66C165}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E4BA8FB4-ACD6-43EB-BB03-14AA0C376F5B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E4BA8FB4-ACD6-43EB-BB03-14AA0C376F5B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E4BA8FB4-ACD6-43EB-BB03-14AA0C376F5B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E4BA8FB4-ACD6-43EB-BB03-14AA0C376F5B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EA7CBA29-CFA8-4945-BB64-CCF25078190E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EA7CBA29-CFA8-4945-BB64-CCF25078190E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EA7CBA29-CFA8-4945-BB64-CCF25078190E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EA7CBA29-CFA8-4945-BB64-CCF25078190E}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+	EndGlobalSection
+	GlobalSection(MonoDevelopProperties) = preSolution
+		Policies = $0
+		$0.TextStylePolicy = $1
+		$1.inheritsSet = null
+		$1.scope = text/x-csharp
+		$0.CSharpFormattingPolicy = $2
+		$2.IndentSwitchBody = True
+		$2.IndentBlocksInsideExpressions = True
+		$2.AnonymousMethodBraceStyle = NextLine
+		$2.PropertyBraceStyle = NextLine
+		$2.PropertyGetBraceStyle = NextLine
+		$2.PropertySetBraceStyle = NextLine
+		$2.EventBraceStyle = NextLine
+		$2.EventAddBraceStyle = NextLine
+		$2.EventRemoveBraceStyle = NextLine
+		$2.StatementBraceStyle = NextLine
+		$2.ElseNewLinePlacement = NewLine
+		$2.CatchNewLinePlacement = NewLine
+		$2.FinallyNewLinePlacement = NewLine
+		$2.WhileNewLinePlacement = DoNotCare
+		$2.ArrayInitializerWrapping = DoNotChange
+		$2.ArrayInitializerBraceStyle = NextLine
+		$2.BeforeMethodDeclarationParentheses = False
+		$2.BeforeMethodCallParentheses = False
+		$2.BeforeConstructorDeclarationParentheses = False
+		$2.NewLineBeforeConstructorInitializerColon = NewLine
+		$2.NewLineAfterConstructorInitializerColon = SameLine
+		$2.BeforeDelegateDeclarationParentheses = False
+		$2.NewParentheses = False
+		$2.SpacesBeforeBrackets = False
+		$2.inheritsSet = Mono
+		$2.inheritsScope = text/x-csharp
+		$2.scope = text/x-csharp
+		$0.TextStylePolicy = $3
+		$3.NoTabsAfterNonTabs = True
+		$3.EolMarker = Windows
+		$3.inheritsSet = VisualStudio
+		$3.inheritsScope = text/plain
+		$3.scope = text/plain
+		$0.TextStylePolicy = $4
+		$4.inheritsSet = null
+		$4.scope = application/config+xml
+		$0.XmlFormattingPolicy = $5
+		$5.inheritsSet = null
+		$5.scope = application/config+xml
+		$0.TextStylePolicy = $6
+		$6.inheritsSet = null
+		$6.scope = application/xml
+		$0.XmlFormattingPolicy = $7
+		$7.inheritsSet = Mono
+		$7.inheritsScope = application/xml
+		$7.scope = application/xml
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
