+++ packages/repositories.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<repositories>
+  <repository path="../SharpFileSystem.Tests/packages.config" />
+  <repository path="../SharpFileSystem.SevenZip/packages.config" />
+  <repository path="../SharpFileSystem.SharpZipLib/packages.config" />
+  <repository path="../SharpFileSystem.Tests/packages.config" />
+</repositories>
